display: JDI: Keep backlight on when shutting down LCD
Currently, the backlight gets turned down to 0 on suspend.
However, on resume, the first light sensor reading takes too long
and the backlight takes several seconds before it starts to ramp up
according to the ambient light.

Keep the backlight level saved so that we can immediately have a
legible screen on resume.

Change-Id: I744d7639f3073bb0e66ed5e1225ffaf10ac35956